+++ src/remap.hpp
@@ -22,7 +22,7 @@ struct Remap {
 
     virtual void update();
 
-    virtual uint8_t get8(uint8_t* data);
+    virtual float get8(uint8_t* data);
     virtual uint16_t get16(uint8_t* data);
     virtual uint32_t get24(uint8_t* data);
     virtual uint32_t get32(uint8_t* data);
